feat: use class cost in the shop fix: -1 skill point at start

+++ src/Application/include/models/Class.hpp
@@ -26,7 +26,7 @@ struct Class {
     float maxHealth;
     float damage;
     float speed;
-    size_t cost;
+    int cost;
     engine::d2::HitboxSolid hitbox;
 
     std::vector<EntityFactory::ID> children;
